Docs and disable debug logging

+++ pc-club-server/Services/SteamService/ISteamService.cs
@@ -4,7 +4,7 @@ namespace pc_club_server.Services.SteamService
 {
     public interface ISteamService
     {
-        string GetGameBannerURI(int gameID);
+        string GetGameBannerURI(long gameID);
         Task<UserGameDataModel> GetUserGames(string steamID64);
     }
 }
